GLOBAL CORS - correct allowedMethods, test separation --> Add, Delete, Get and Update, create testConfig, moving URL to application.properties
