+++ rest-app/src/main/resources/application.properties
@@ -2,6 +2,7 @@
 # SERVER CONFIG
 #------------------------------------------------------
 server.port=8090
+url = http://localhost:8090/order
 #------------------------------------------------------
 
 #------------------------------------------------------
@@ -29,10 +30,7 @@ logging.level.org.hibernate.type=trace
 #------------------------------------------------------
 cors.addMapping=/**
 cors.allowedOrigins=http://localhost:4200
-cors.allowedMethodsGet=GET
-cors.allowedMethodsPost=POST
-cors.allowedMethodsPut=PUT
-cors.allowedMethodsDelete=DELETE
+cors.allowedMethods=GET,POST,PUT,DELETE
 cors.allowedHeaders=*
 cors.allowCredentials=false
 cors.maxAge=3600
