--- rest-app/src/test/java/com/epam/repair/cucumber/CucumberRunnerTest.java
@@ -1,10 +1,0 @@
-package com.epam.repair.cucumber;
-
-import io.cucumber.junit.Cucumber;
-import io.cucumber.junit.CucumberOptions;
-import org.junit.runner.RunWith;
-
-@RunWith(Cucumber.class)
-@CucumberOptions(plugin = "pretty", features = "src/test/resources/features")
-public class CucumberRunnerTest {
-}
