+++ rest-app/src/test/resources/features/editOrder.feature
@@ -8,9 +8,9 @@ Feature: Edit order
   Scenario: Validation. UNSUCCESSFULLY order edit
     Given User need to edit the client first name in a order.
     When Find the order.
-    Then Write the client first name "AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA".
+    Then Write the client lastName >= than 41 characters.
 
   Scenario: Validation. UNSUCCESSFULLY order edit
     Given User need to edit the client first name in a order.
     When Find the order.
-    Then Write the customer first name "".
+    Then Write the client lastName than 0 characters.
